refactor(openapi): unify duplicated tag schemas into shared components
Tag definitions were duplicated and inconsistent across firewall rules,
public keys, devices, and container endpoints. Introduce shared
components: tagName (name string), tagNameArray (list of names), and a
tagNamePath path parameter, reusing the existing tag schema for full
objects.

Split publicKeyFilter into request and response variants: the request
takes tag names while the response always emits full tag objects, so a
single shared schema could not describe both.

Adapt the console UI to the regenerated client. The public key response
filter is now a flat object, so normalizePublicKey is dropped and
PublicKeyResponse is consumed directly, mapping tag objects to names in
the key drawer. The firewall request filter types tags as plain name
strings, so the rule drawer no longer wraps them in objects.
