--- openapi/spec/components/parameters/path/deviceTagPath.yaml
@@ -1,6 +1,0 @@
-name: tag
-description: Device's tag name
-schema:
-  $ref: ../../schemas/tag.yaml
-required: true
-in: path
