+++ openapi/spec/components/schemas/firewallRulesResponse.yaml
@@ -19,26 +19,21 @@ properties:
     example: true
   filter:
     description: Firewall rule's filter
-    oneOf:
-      - type: object
-        properties:
-          hostname:
-            description: Firewall rule's hostname
-            type: string
-            example: .*
-        required:
-          - hostname
-      - type: object
-        properties:
-          tags:
-            description: Firewall's rule tags
-            type: array
-            items:
-              $ref: tag.yaml
-            minItems: 1
-            maxItems: 3
-        required:
-          - tags
+    type: object
+    properties:
+      hostname:
+        description: Firewall rule's hostname
+        type: string
+        example: .*
+      tags:
+        description: Firewall's rule tags
+        type: array
+        items:
+          $ref: tag.yaml
+        minItems: 0
+        maxItems: 3
+    required:
+      - tags
   priority:
     description: Firewall rule's priority
     type: integer
