--- openapi/spec/components/schemas/publicKeyFilter.yaml
@@ -1,31 +1,0 @@
-description: |
-  Public key's filter rule.
-
-
-  The `filter`` rule defines how if the public key is valid to a device.
-
-  - When `hostname` object is set, the public key will be used in a device what matches with hostname.
-  - When `tags` object is set, it matches the device what contains at least one of that tags.
-oneOf:
-  - type: object
-    properties:
-      hostname:
-        description: Public key's regex hostname.
-        type: string
-        example: .*
-    required:
-      - hostname
-  - type: object
-    properties:
-      tags:
-        description: Public key's tags.
-        type: array
-        items:
-          type: string
-          minLength: 3
-          maxLength: 255
-        minItems: 1
-        maxItems: 3
-        uniqueItems: true
-    required:
-      - tags
