+++ openapi/spec/paths/api@namespaces@{tenant}@containers@{uid}@tags@{name}.yaml
@@ -12,14 +12,7 @@ post:
   parameters:
     - $ref: ../components/parameters/path/namespaceTenantIDPath.yaml
     - $ref: ../components/parameters/path/deviceUIDPath.yaml
-    - name: name
-      description: Tag name to associate
-      schema:
-        type: string
-        pattern: ^[a-zA-Z0-9]+$
-        example: prod
-      required: true
-      in: path
+    - $ref: ../components/parameters/path/tagNamePath.yaml
   responses:
     '204':
       description: Tag successfully associated with container
@@ -47,14 +40,7 @@ delete:
   parameters:
     - $ref: ../components/parameters/path/namespaceTenantIDPath.yaml
     - $ref: ../components/parameters/path/deviceUIDPath.yaml
-    - name: name
-      description: Tag name to remove
-      schema:
-        type: string
-        pattern: ^[a-zA-Z0-9]+$
-        example: prod
-      required: true
-      in: path
+    - $ref: ../components/parameters/path/tagNamePath.yaml
   responses:
     '204':
       description: Tag successfully removed from container
