+++ openapi/spec/paths/api@tags@{name}.yaml
@@ -9,14 +9,7 @@ patch:
     - jwt: []
     - api-key: []
   parameters:
-    - name: name
-      description: Current tag name
-      schema:
-        type: string
-        pattern: ^[a-zA-Z0-9]+$
-        example: prod
-      required: true
-      in: path
+    - $ref: ../components/parameters/path/tagNamePath.yaml
   requestBody:
     required: true
     content:
@@ -27,12 +20,7 @@ patch:
             - name
           properties:
             name:
-              type: string
-              minLength: 3
-              maxLength: 255
-              pattern: ^[a-zA-Z0-9]+$
-              example: dev
-              description: New tag name
+              $ref: ../components/schemas/tagName.yaml
   responses:
     '200':
       description: Tag successfully updated
@@ -63,14 +51,7 @@ delete:
     - jwt: []
     - api-key: []
   parameters:
-    - name: name
-      description: Tag name to delete
-      schema:
-        type: string
-        pattern: ^[a-zA-Z0-9]+$
-        example: prod
-      required: true
-      in: path
+    - $ref: ../components/parameters/path/tagNamePath.yaml
   responses:
     '204':
       description: Tag successfully deleted
